Renamed some files/components for consistency
